moving const.ts to Utils/
