+++ src/pages/modules/MainPhase/MainPhase.test.js
@@ -3,7 +3,7 @@ import { render, fireEvent, act, waitForElement, wait } from '@testing-library/r
 import '@testing-library/jest-dom/extend-expect';
 import fetch from 'jest-fetch-mock';
 import MainPhase from './MainPhase';
-import { SEARCH_PHASE_MOCK_RESPONSE, SEARCH_PHASE_MOCK_RESPONSE_PARTIAL } from '../util/const';
+import { SEARCH_PHASE_MOCK_RESPONSE, SEARCH_PHASE_MOCK_RESPONSE_PARTIAL } from 'Utils/const';
 import GlobalContext from '../util/GlobalContext.tsx';
 
 describe('MainPhase.js', () => {
