+++ src/pages/modules/MainPhase/submodules/DataForm/DataForm.test.js
@@ -7,7 +7,7 @@ import {
 import '@testing-library/jest-dom/extend-expect';
 import fetch from 'jest-fetch-mock';
 import DataForm from './DataForm';
-import { MEDIA_STATUS_COLORS, ANILIST_BASE_URL } from '../../../util/const';
+import { MEDIA_STATUS_COLORS, ANILIST_BASE_URL } from 'Utils/const';
 import GlobalContext from '../../../util/GlobalContext.tsx';
 
 const setup = ({ presetProgress = 0, presetScore = 0 } = {}) => {
