+++ src/pages/modules/MainPhase/submodules/SearchPhase/SearchPhase.test.js
@@ -3,7 +3,7 @@ import { render, fireEvent, act } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import fetch from 'jest-fetch-mock';
 import SearchPhase from './SearchPhase.tsx';
-import { SEARCH_PHASE_MOCK_RESPONSE, SEARCH_PHASE_MOCK_RESPONSE_PARTIAL } from '../../../util/const';
+import { SEARCH_PHASE_MOCK_RESPONSE, SEARCH_PHASE_MOCK_RESPONSE_PARTIAL } from 'Utils/const';
 import GlobalContext from '../../../util/GlobalContext.tsx';
 
 const setup = () => {
